Changed GSYM header to specify string table via file  offset and size.
This allows the GSYM header to be easily skipped while being parsed in a native parser. It also allows the GSYM file format to not require an object file format like mach-o, ELF or COFF, as the file can be a stand alone file since it doesn't require the string table to be a section in an object file.

Fixed a few other things:
- Make sure .gsym section is 16 byte aligned in ELF files
- Corrreclty set the ELF section header for the GSYM string table to be SHT_STRTAB.
- Updated README.md with new file format changes
- Removed documentation in gsym.py since this data is in README.md and the info was duplicated and becoming out of date